+++ src/core/db/EntryTransaction.ts
@@ -20,7 +20,7 @@ import {SourceTransaction} from '../source/Source.js'
 import type {ReadonlyTree} from '../source/Tree.js'
 import {assert} from '../util/Assert.js'
 import {type CommitChange, commitChanges} from './CommitRequest.js'
-import type {EntryIndex} from './EntryGraph.js'
+import type {EntryIndex} from './EntryIndex.js'
 import type {
   ArchiveMutation,
   CreateMutation,
