+++ src/core/db/LocalDB.ts
@@ -8,7 +8,7 @@ import {MemorySource} from '../source/MemorySource.js'
 import type {Source} from '../source/Source.js'
 import {syncWith} from '../source/Source.js'
 import {type CommitRequest, sourceChanges} from './CommitRequest.js'
-import {EntryIndex} from './EntryGraph.js'
+import {EntryIndex} from './EntryIndex.js'
 import {EntryResolver} from './EntryResolver.js'
 import {EntryTransaction} from './EntryTransaction.js'
 import type {Mutation} from './Mutation.js'
